Move grammar package
